Рефактор №2

+++ src/main/java/ru/majordomo/hms/personmgr/repository/AccountPromocodeRepository.java
@@ -19,7 +19,7 @@ public interface AccountPromocodeRepository extends MongoRepository<AccountPromo
     AccountPromocode findOneByPromocodeId(@Param("promocodeId") String promocodeId);
     AccountPromocode findByPromocodeIdAndOwnedByAccount(@Param("promocodeId") String promocodeId, @Param("ownedByAccount") boolean ownedByAccount);
     AccountPromocode findOneByPersonalAccountIdAndOwnedByAccount(@Param("personalAccountId") String personalAccountId, @Param("ownedByAccount") boolean ownedByAccount);
-    Integer countByPersonalAccountIdAndOwnedByAccount(@Param("personalAccountId") String personalAccountId, @Param("ownedByAccount") boolean ownedByAccount);
+    Long countByPersonalAccountIdAndOwnedByAccount(@Param("personalAccountId") String personalAccountId, @Param("ownedByAccount") boolean ownedByAccount);
     List<AccountPromocode> findByOwnedByAccount(@Param("ownedByAccount") boolean ownedByAccount);
 
     @RestResource(path = "findListByPersonalAccountId", rel = "findListByPersonalAccountId")
